Improvements to text on 'Verify payment' page
Shortened/simplified some text.

Changed placeholders from the word "transaction" to the shorthand "tx". Since the user already sees the word transaction to the left of the field, they understand that they mean the same thing, which is important since the History page, where most will get their tx IDs and tx keys, use "tx" instead of "transaction".

Also the placeholders were changed from "Transaction ID here" to "Paste transaction ID". The text to the left of the field already is an indicator of what the box is for, so the placeholder instead acts as an action for the user to take (paste!).