Adjectives or nouns for Languages
I thought it was agreed some time back that the language names would be adjectives in their own languages, but it does not seem that this has been effected. In the meantime, english adjectives can suffice. What does *not* make sense is a mix of adjectives and nouns. "Choose your language: English, Russia, or Brasil" sounds potato.