RemoteNodeEdit: fix treatment of IPv6 addresses
Now RemoteNodeEdit items check if the host is an IPv6 address and fixes
the return of `getAddress()`, for example, if it's an IPv6 address, it
returns [2001:db8::1]:18081 instead of 2001:db8::1:18081 which is
ambiguous and also showed up incorrectly on the GUI.

Previously the user had to enter the address inside square brackets in
order to connect to an IPv6 remote node (and it succeed, ofc) however
the GUI didn't parsed it correctly.

These bugs were found during the test of the Locha Mesh network (which
uses `fc00::/16` unique local addresses) as can be seen on the
[demonstration].

[demonstration]: https://twitter.com/Locha_io/status/1276930442030133248

The incorrect parsing is showed on issue #2971

The `getAddressParts()` will try to guess, depending on the input, how
the address is structured, if it is:

1. IPv6 with a port.
2. IPv6 without a port.
3. IPv4/DNS with a port.
4. IPv4/DNS without a port.
5. Treat anything else as an error and use the entire string as the
host.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>